requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.1
Jinja2==3.1.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
reportlab==4.2.5
email-validator==2.2.0
